Review #1097 useful tooltip
