+++ src/components/Tooltip/index.tsx
@@ -1,0 +1,21 @@
+import React from 'react'
+import ReactTooltip from 'react-tooltip'
+import styled from 'styled-components'
+
+const StyledTooltip = styled(ReactTooltip)`
+  opacity: 1 !important;
+  z-index: 9999 !important;
+`
+
+const Tooltip: React.FC = props => {
+  return (
+    <StyledTooltip
+      event="mouseenter focus"
+      eventOff="mouseleave blur"
+      effect="solid"
+      {...props}
+    />
+  )
+}
+
+export default Tooltip
